+++ R/parse_spliceai.R
@@ -47,8 +47,8 @@ parse_spliceai <- function(vcf_file){
     mutate(
 
       # replace missing data with "." with NA
-      across(starts_with("DS_"), na_if, "."),
-      across(starts_with("DP_"), na_if, "."),
+      across(starts_with("DS_"), \(x) na_if(x, ".")),
+      across(starts_with("DP_"), \(x) na_if(x, ".")),
 
       # convert scores as numeric and positions as integers
       across(starts_with("DS_"), as.numeric),
@@ -59,3 +59,69 @@ parse_spliceai <- function(vcf_file){
     left_join(splice_df, by = "Key")
 
 }
+
+
+
+#' Parse VCF output file from SpliceAI with -T flag as table
+#'
+#' requires the vcfR package.
+#'
+#' @param vcf_file path to a single VCF file, the output of spliceAI
+#' @return a [tibble][tibble::tibble-package] with one row per variant
+#' annotation. Each input variant can have multiple annotations.
+#'
+#' @examples
+#'
+#' spliceai_file <- system.file("extdata", "spliceai_thresh_output.vcf", package = "splice2neo")
+#' parse_spliceai_thresh(spliceai_file)
+#'
+#' @seealso \code{\link{parse_spliceai}}, \code{\link{format_spliceai_thresh}}, \code{\link{annotate_mut_effect}}
+#' @import dplyr stringr tidyr
+#' @export
+parse_spliceai_thresh <- function(vcf_file){
+
+  # read in vcf file
+  vcf <- vcfR::read.vcfR(vcf_file, verbose = FALSE)
+
+  # get fixed fields for variants
+  if (nrow(vcf) > 1){
+    fix_df <-  vcfR::getFIX(vcf) %>%
+      tibble::as_tibble() %>%
+      mutate(Key = row_number())
+  } else {
+    fix_df <-  vcfR::getFIX(vcf) %>%
+      tibble::as_tibble_row() %>%
+      mutate(Key = row_number())
+  }
+
+  # get effect annotations for all mutations (separate alleles on same position)
+  splice_df <- vcf %>%
+    vcfR::extract_info_tidy(info_fields = c("SpliceAI")) %>%
+    mutate(fields = SpliceAI %>% stringr::str_split(",")) %>%
+    dplyr::select(-SpliceAI) %>%
+    tidyr::unnest(fields) %>%
+    dplyr::filter(!is.na(fields))
+
+  # separate multiple effect predictions per mutation (unique alleles)
+  splice_df <- splice_df %>%
+    tidyr::separate(fields,
+                    into = c("ALLELE", "SYMBOL", "splice_sites"),
+                    sep = "\\|",
+                    extra = "merge") %>%
+    mutate(fields = splice_sites %>% stringr::str_split("\\|")) %>%
+    dplyr::select(-splice_sites) %>%
+    tidyr::unnest(fields) %>%
+    # replace empty fields with NA
+    mutate(fields = na_if(fields, "")) %>%
+    tidyr::separate(fields, into = c("effect", "score", "pos_rel"), sep = ":") %>%
+    mutate(
+      score = as.numeric(score),
+      pos_rel = as.integer(pos_rel),
+      effect = as.factor(effect)
+    )
+
+  fix_df %>%
+    left_join(splice_df, by = "Key")
+
+}
+