+++ NEWS.md
@@ -1,3 +1,35 @@
+# splice2neo 0.6.7
+
+## Added:
+
+ - Functionality from splice2neo_neoants:
+    - CI-SpliceAI parsing and formatting functions with example data and tests
+    - IRFinder parsing and filtering functions with example data and tests
+    - StringTie parsing and formatting functions
+    - SUPPA2 parsing and formatting functions
+    - STAR parsing and formatting functions with example data and tests
+    - RegTools parsing and formatting functions with example data and tests
+ - add test for empty CI-SpliceAI input file
+ - add more global variables to avoid notes in R CMD check
+
+## Changed:
+
+ - Update CI/CD pipeline for less verbosity during R pkg installations
+ - Return single-row data.frame for empty input in `annotate_mut_effect()`
+ - include 'number_of_supporting_reads` when parsing SplAdder output
+ - generalize `generate_combined_dataset()` to multiple inputs
+ - Update README with list of supported tools
+ - add option consider_intron_retention to `annotate_mut_effect()`
+ - adjust syntax for dplyr >= 1.1.1
+ - remove rarely used pkgs from Imports
+ - simplify code in parse_spliceai_thresh()
+ - minor spell check fixes and doc updates
+
+## Fixed: 
+
+ - fix warnings due to NA's in CI-SpliceAI parsing function
+ - fix use of `generate_combined_dataset()` in vignette
+
 # splice2neo 0.6.6
 
 * add vignette with example workflow
@@ -68,7 +100,7 @@
 
 # splice2neo 0.5.0
 
-* intron rentetion events are now supported by `add_context_seq()`. the resulting context sequence covers the complete intron instead of the exon/intron boundary only. Instead of a the junction position in the cts_seq, the positions are given in form of an interval in the `cts_junc_pos`  column for intron retentions. (0,start_IR, end_IR, end_cts) 
+* intron retention events are now supported by `add_context_seq()`. the resulting context sequence covers the complete intron instead of the exon/intron boundary only. Instead of a the junction position in the cts_seq, the positions are given in form of an interval in the `cts_junc_pos`  column for intron retentions. (0,start_IR, end_IR, end_cts) 
 * `add_peptide()` was adjusted for intron retentions
 * more tests were added for several functions
 * fix small bug in `annotate_spliceai_junction()` that led to annotation with same transcript 
@@ -83,7 +115,7 @@
 
 # splice2neo 0.4.0
 
-* Leafcutterr: The strand information while leafcutter parsing is now retrieved from the cluster id in the count table.
+* Leafcutter: The strand information while leafcutter parsing is now retrieved from the cluster id in the count table.
 * Spladder: previous code changes missed to consider the event type in generation of the junc id for alternative splice sites. This is fixed now.
 * More tests were added
 * An old bug in spladder_transform_mutex_exon was fixed